+++ Front/src/components/drivings/Drivings.tsx
@@ -6,277 +6,19 @@ import { userAccess, userToken } from '../login/loginSlice';
 import { getOrdersAsync, orderEndedAsync, ordersSelector } from '../orders/OrdersSlice';
 import { allDrivesSelector, drivesSelector, endDriveAsync, getAllDrivesAsync, getDrivesAsync, startDriveAsync } from './drivesSlicer';
 import { DriveModel } from '../../models/Drive';
+import PreviousDrives from './PreviousDrives';
+import ActiveDrive from './ActiveDrive';
 
 
 export function Drivings() {
-  const dispatch = useAppDispatch()
-  const token = useAppSelector(userAccess)
-  const decoded: any = jwt_decode(token)
-  const [isRunning, setIsRunning] = useState(false);
-  const drives = useAppSelector(drivesSelector)
-  const allDrives = useAppSelector(allDrivesSelector)
-  const orders = useAppSelector(ordersSelector)
-  const [activeOrder, setactiveOrder] = useState<OrderModel | null>(null)
-  const [refreshFlag, setrefreshFlag] = useState(false)
-  const [startKilometer, setstartKilometer] = useState("")
-  const [endKilometer, setendKilometer] = useState("")
-  const [changeKilometer, setchangeKilometer] = useState(false)
-  const [startSelectedFile1, setstartSelectedFile1] = useState<File | null>(null)
-  const [startSelectedFile2, setstartSelectedFile2] = useState<File | null>(null)
-  const [startSelectedFile3, setstartSelectedFile3] = useState<File | null>(null)
-  const [endSelectedFile1, setendSelectedFile1] = useState<File | null>(null)
-  const [endSelectedFile2, setendSelectedFile2] = useState<File | null>(null)
-  const [endSelectedFile3, setendSelectedFile3] = useState<File | null>(null)
-  const [activeDrive, setactiveDrive] = useState<DriveModel | null>(null)
-  const [comments, setcomments] = useState("")
-  const [activeDriveFlag, setactiveDriveFlag] = useState(false)
-  const [selectedDrive, setselectedDrive] = useState<DriveModel | null>(null)
-
-  // Handles image 1 upload
-  const handleFile1Change = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setstartSelectedFile1(e.target!.files![0])
-  };
-  // Handles image 2 upload
-  const handleFile2Change = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setstartSelectedFile2(e.target!.files![0])
-  };
-  // Handles image 3 upload
-  const handleFile3Change = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setstartSelectedFile3(e.target!.files![0])
-  };
-
-  // Handles image 1 upload
-  const handleendFile1Change = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setendSelectedFile1(e.target!.files![0])
-  };
-  // Handles image 2 upload
-  const handleendFile2Change = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setendSelectedFile2(e.target!.files![0])
-  };
-  // Handles image 3 upload
-  const handleendFile3Change = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setendSelectedFile3(e.target!.files![0])
-  };
-
-
-  // Detects the active order by the date of today
-  const handleCurrentOrder = () => {
-    const today = new Date()
-    const currentOrder = orders.find(order => new Date(order.fromDate).getTime() <= today.getTime() && new Date(order.toDate).getTime() >= today.getTime() && order.ended === false)
-    setactiveOrder(currentOrder!)
-    setrefreshFlag(!refreshFlag)
-  }
-  // Gets the end kilometer of the car of the active order, according to the report of the last drive
-  const handleStartKilometer = () => {
-    const lastDrive = allDrives.filter(drive => drive.endKilometer && drive.car_name === activeOrder?.car_name).pop()
-    lastDrive ? setstartKilometer(String(lastDrive?.endKilometer)) : setstartKilometer("")
-  }
-
-  // Calls the server with GET methods When the app goes live
-  useEffect(() => {
-    dispatch(getDrivesAsync(token))
-    dispatch(getOrdersAsync(token))
-    dispatch(getAllDrivesAsync(token))
-  }, [token])
-
-  // Checks if there is an active drive when app is loaded
-  useEffect(() => {
-    const startStopBtn = document.querySelector('.round') as HTMLButtonElement;
-    if (localStorage.hasOwnProperty('isRunning')) {
-      startStopBtn.textContent = 'Stop';
-      startStopBtn.className = "round redBtn";
-      setIsRunning(true)
-    }
-  }, [])
-
-  // Checks if there is an active drive when app is loaded
-  useEffect(() => {
-    if (localStorage.hasOwnProperty('activeDrive')) {
-      setactiveDrive(drives[drives.length - 1])
-    }
-    console.log(activeOrder)
-  }, [])
-
-  // Checks if there is an order today
-  useEffect(() => {
-    handleCurrentOrder()
-  }, [orders.length, token])
-
-  // Handles the active drive change
-  useEffect(() => {
-    dispatch(getDrivesAsync(token))
-    setactiveDriveFlag(!activeDriveFlag)
-  }, [drives.length])
-
-
-  useEffect(() => {
-    isRunning && setactiveDrive(drives[drives.length - 1])
-  }, [activeDriveFlag])
-
-  // Gets the kilometer of the active order, if exists.
-  useEffect(() => {
-    handleStartKilometer()
-  }, [refreshFlag, allDrives.length, token])
-
-
-  const handleButtonClick = () => {
-    const startStopBtn = document.querySelector('.round') as HTMLButtonElement;
-    setIsRunning(!isRunning);
-    if (!isRunning) {
-      localStorage.setItem('isRunning', "true")
-      localStorage.setItem('activeDrive', "true")
-      dispatch(startDriveAsync({
-        token: token, drive: {
-          user: decoded.user_id,
-          order: activeOrder?.id,
-          startDate: new Date(),
-          startKilometer: startKilometer,
-          startImg1: startSelectedFile1,
-          startImg2: startSelectedFile2,
-          startImg3: startSelectedFile3,
-        }
-      }))
-
-      startStopBtn.textContent = 'Stop';
-      startStopBtn.className = "round redBtn";
-    } else {
-      dispatch(endDriveAsync({
-        token: token, drive: {
-          id: activeDrive?.id,
-          endDate: new Date(),
-          comments: comments,
-          endKilometer: endKilometer,
-          endImg1: endSelectedFile1,
-          endImg2: endSelectedFile2,
-          endImg3: endSelectedFile3,
-        }
-      }))
-      dispatch(orderEndedAsync({
-        token: token,
-        id: activeOrder?.id!
-      }))
-      setactiveDrive(null)
-      setactiveOrder(null)
-      localStorage.removeItem('isRunning')
-      localStorage.removeItem('activeDrive')
-      startStopBtn.textContent = 'Start';
-      startStopBtn.className = "round greenBtn";
-    }
-  };
 
   return (
     <div >
       <h1>נסיעות קודמות</h1>
-      <div style={{ display: 'grid', gridTemplateColumns: 'repeat(auto-fill, minmax(200px, 1fr))', gap: '.25rem', gridAutoRows: 'minmax(160px, auto)' }}>
-
-        {drives.filter(drive => drive.endDate).map(drive =>
-
-          <div key={drive.id} style={{ borderRadius: '5px', border: '2px solid rgb(0, 0, 0)', padding: '.5rem', margin: "10px" }}>
-
-            מכונית: {drive.car_name}<br />
-            <img src={`http://127.0.0.1:8000${drive.car_image}`} style={{ width: '150px', height: '100px' }} alt={drive.car_name} /><br /><br />
-            {drive.startDate!.toString().slice(0, 10) === drive.endDate!.toString().slice(0, 10) ?
-              <div>
-                בתאריך: {drive.startDate!.toString().slice(0, 10)}<br />
-              </div> :
-              <div>
-                מתאריך: {drive.startDate!.toString().slice(0, 10)}<br />
-                עד תאריך: {drive.endDate!.toString().slice(0, 10)}<br />
-              </div>
-            }
-
-            משעה: {drive.startDate!.toString().slice(11, 16)}<br />
-            עד שעה: {drive.endDate!.toString().slice(11, 16)}<br />
-            קילומטראז' התחלתי: {Number(drive.startKilometer).toLocaleString()}<br />
-            קילומטראז' סופי: {Number(drive.endKilometer).toLocaleString()}<br />
-            הערות: {drive.comments ? drive.comments : 'אין הערות'}<br />
-            <button onClick={() => setselectedDrive(drive)}>לצפייה בכל הפרטים</button>
-          </div>)}
-      </div>
+      <PreviousDrives/>
       <div>
-        {isRunning ?
-          <div>
-            <h1>נסיעה פעילה</h1><hr />
-            הנסיעה התחילה
-            {activeDrive?.id}
-
-            <img src={`http://127.0.0.1:8000${activeOrder?.car_image}`} style={{ width: '150px', height: '100px' }} alt={activeDrive?.car_name} /><br />
-            בשעה: {activeDrive?.startDate!.toString().slice(11, 16)}<br />
-            קילומטראז': <input onChange={(e) => setendKilometer(e.target.value)} value={endKilometer} /><br />
-
-            <input type='file' onChange={handleendFile1Change} /><br />
-            {startSelectedFile1 &&
-              <div>
-                <input type='file' onChange={handleendFile2Change} /><br />
-              </div>}
-            {startSelectedFile2 &&
-              <div>
-                <input type='file' onChange={handleendFile3Change} /><br />
-              </div>}
-            הערות: <input onChange={(e) => setcomments(e.target.value)} />
-                
-            <div className="d-flex justify-content-center">
-                  <button className="round greenBtn" onClick={handleButtonClick}>Stop</button>
-                </div>
-
-          </div> :
-          <div>
-            {(activeOrder && !activeOrder?.ended) &&
-              <div>
-                <hr />
-                <h3>הזמנה פעילה</h3>
-                <div> {activeOrder.car_name}</div><br />
-                <div>{activeOrder.ended}</div>
-                <img src={`http://127.0.0.1:8000${activeOrder.car_image}`} style={{ width: '150px', height: '100px' }} alt={activeOrder.car_name} /><br />
-                {activeOrder.isAllDay ?
-                  <div>
-                    כל היום
-                  </div> :
-                  <div>
-                    משעה: {activeOrder.fromDate!.toString().slice(11, 16)}<br />
-                    עד שעה: {activeOrder.toDate!.toString().slice(11, 16)}<br />
-                  </div>
-                }
-                {startKilometer ?
-                  <div>
-                    קילומטראז': <input onChange={(e) => setstartKilometer(e.target.value)} value={startKilometer} disabled={!changeKilometer} />
-                    דרוש שינוי? <input type={'checkbox'} onChange={() => setchangeKilometer(!changeKilometer)} /><br />
-                  </div> :
-                  <div>
-                    קילומטראז': <input onChange={(e) => setstartKilometer(e.target.value)} value={startKilometer} />
-                  </div>
-                }
-                <input type='file' onChange={handleFile1Change} /><br />
-                {startSelectedFile1 &&
-                  <div>
-                    <input type='file' onChange={handleFile2Change} /><br />
-                  </div>}
-                {startSelectedFile2 &&
-                  <div>
-                    <input type='file' onChange={handleFile3Change} /><br />
-                  </div>}
-
-
-                <div className="d-flex justify-content-center">
-                  <button className="round greenBtn" onClick={handleButtonClick}>Start</button>
-                </div>
-
-              </div>
-            }
-
-          </div>
-        }
-
+        <ActiveDrive/>
       </div>
-      {/* {selectedDrive &&
-        <div style={{ position: "fixed", top: "0", left: "0", width: "100%", height: "100vh", backgroundColor: "rgba(0,0,0,0.2)", display: "flex", justifyContent: "center", alignItems: "center" }}>
-          <div style={{ position: "relative", padding: "32px", width: "400px", height: "300px", maxWidth: "640px", backgroundColor: "white", border: "2px solid black", borderRadius: "5px" }}>
-            <img src={`http://127.0.0.1:8000${selectedCar.image}`} style={{ width: '150px', height: '100px' }} alt={selectedCar.model} /><br /><br />
-            יעד נסיעה: <input onChange={(e) => setdestination(e.target.value)} />
-            <button onClick={() => handleOrder()}>הזמן</button>
-          </div>
-        </div>
-      } */}
 
     </div>
   );
